requirements.txt
Flask==3.1.2
pandas
pytesseract
Pillow
pdf2image
requests
google-genai
